# specs/001-physical-ai-textbook/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Physical AI Textbook API
  description: RAG Chatbot, Authentication, Personalization, and Translation APIs
  version: 1.0.0
  contact:
    name: Physical AI Team
servers:
  - url: https://api.physicalairobotics.com/v1
    description: Production
  - url: http://localhost:8000/v1
    description: Development
...
paths:
...
  /user/preferences:
    get:
      tags: [User]
      summary: Get user preferences
      operationId: getPreferences
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferences'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags: [User]
      summary: Update user preferences
      operationId: updatePreferences
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPreferences'
      responses:
        '200':
          description: Preferences updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferences'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /user/questionnaire:
    post:
      tags: [User]
      summary: Submit background questionnaire
      operationId: submitQuestionnaire
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionnaireSubmission'
      responses:
        '200':
          description: Questionnaire submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferences'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /user/progress:
    get:
      tags: [User]
      summary: Get reading progress
      operationId: getProgress
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Reading progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadingProgress'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [User]
      summary: Update reading progress
      operationId: updateProgress
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProgressUpdate'
      responses:
        '200':
          description: Progress updated

  # ============ TRANSLATION ENDPOINTS ============
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserPreferences:
      type: object
      properties:
        experience_level:
          type: string
          enum: [beginner, intermediate, advanced]
        programming_langs:
          type: array
          items:
            type: string
        robotics_experience:
          type: boolean
        hardware_available:
          type: array
          items:
            type: string
        learning_goals:
          type: array
          items:
            type: string
        preferred_language:
          type: string
          enum: [en]  # Urdu deferred to future phase (2025-12-20)
        theme:
          type: string
          enum: [light, dark, system]
        font_size:
          type: string
          enum: [small, medium, large]
    QuestionnaireSubmission:
      type: object
      required: [experience_level]
      properties:
        experience_level:
          type: string
          enum: [beginner, intermediate, advanced]
        programming_langs:
          type: array
          items:
            type: string
        robotics_experience:
          type: boolean
        hardware_available:
          type: array
          items:
            type: string
        learning_goals:
          type: array
          items:
            type: string
    ReadingProgress:
      type: object
      properties:
        current_chapter:
          type: integer
        completed_chapters:
          type: array
          items:
            type: integer
        completion_percentage:
          type: number
          format: float
        bookmarks:
          type: array
          items:
            $ref: '#/components/schemas/Bookmark'
        last_read_at:
          type: string
          format: date-time
    Bookmark:
      type: object
      properties:
        chapter:
          type: integer
        section:
          type: string
        note:
          type: string
        created_at:
          type: string
          format: date-time
    ProgressUpdate:
      type: object
      properties:
        chapter_completed:
          type: integer
          minimum: 1
          maximum: 14
        current_chapter:
          type: integer
          minimum: 1
          maximum: 14
        bookmark:
          $ref: '#/components/schemas/Bookmark'

    # ---- Translation Schemas ----
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...